+++ changelog.md
@@ -8,6 +8,7 @@
 - Added plotFormats option
 - tree.add forward the kwargs to all the flowsteps
 - Plots have better handling of the `log` and `density` options. `density` can be `axis` or `counts`
+- implemented processPattern option
 
 fix:
 - Added safeguards against empty regions (0 events)